refactor: 구조 분해
