--- src/components/project-detail/ProjectDetailSection/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ProjectDetailSection';
